minion_sys

Function run_minion

pub fn run_minion(
    model: Model,
    callback: Callback<'_>,
) -> Result<SolverContext, MinionError>

Run Minion on the given Model.

The given callback is ran whenever a new solution set is found.

Returns a SolverContext on success, which can be used to query run statistics via SolverContext::get_from_table.